Revert "Merge remote-tracking branch 'qo0-base/master'"
This reverts commit b6516183ce8899a09cb99943b1cb8fe29c084176, reversing
changes made to 8c8082c3f537487cee7386284b0808ba68b1c5fa.

+++ base/utilities/logging.h
@@ -45,7 +45,7 @@ namespace L
 	/* console write stream */
 	inline FILE*			pStream;
 	/* current color of console text */
-	inline std::uint16_t	wConsoleColor = FOREGROUND_WHITE;
+	inline std::uint8_t		wConsoleColor = FOREGROUND_WHITE;
 	/* current file used for file-logging */
 	inline std::ofstream	ofsFile;
 
@@ -58,7 +58,7 @@ namespace L
 	void Print(std::string_view szText);
 
 	/* set given color to console */
-	inline void PushConsoleColor(const std::uint16_t wColor)
+	inline void PushConsoleColor(const std::uint8_t wColor)
 	{
 		wConsoleColor = wColor;
 	}
